Rereleased ZDK 1.0.2
Fixed JsElem helper function return value.

+++ V3.0/ZDK/src/nl/zeesoft/zdk/json/JsElem.java
@@ -171,7 +171,7 @@ public class JsElem {
 		return getChildLong(name,0L);
 	}
 
-	public long getChildInt(String name) {
+	public int getChildInt(String name) {
 		return getChildInt(name,0);
 	}
 }
